refactor: JWT 페이로드 클레임 enum으로 관리, redis key 구분자 규칙 변경
- JWT 페이로드 클레임이 추후 확장될 가능성이 있을 것 같아 enum으로 관리하도록 변경했습니다.
- redis 키 구분자 규칙을 기존 :: 에서 : 으로 변경 했습니다.
