+++ build.gradle
@@ -45,10 +45,10 @@ dependencies {
 	implementation 'org.apache.commons:commons-lang3:3.14.0'
 
 	compileOnly 'org.projectlombok:lombok'
-	testCompileOnly 'org.projectlombok:lombok'
-	runtimeOnly 'com.mysql:mysql-connector-j'
 	annotationProcessor 'org.projectlombok:lombok'
+	testCompileOnly 'org.projectlombok:lombok'
 	testAnnotationProcessor 'org.projectlombok:lombok'
+	runtimeOnly 'com.mysql:mysql-connector-j'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 
@@ -123,4 +123,8 @@ bootJar {
     from("${generateSwaggerUISample.outputDir}") {
         into 'static/docs'
     }
+}
+
+jar {
+	enabled = false
 }
