+++ src/main/java/com/whatpl/global/jwt/WhatplClaim.java
@@ -1,0 +1,13 @@
+package com.whatpl.global.jwt;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum WhatplClaim {
+
+    NICKNAME("nick");
+
+    private final String key;
+}
